+++ app/pages/browse/index.tsx
@@ -3,6 +3,7 @@ import { GetServerSideProps } from "next";
 import { SelectDatasetStep } from "@/browse/ui/select-dataset-step";
 import { AppLayout } from "@/components/layout";
 import { ConfiguratorStateProvider } from "@/configurator/configurator-state";
+import { useIsEmbedded } from "@/hooks/useIsEmbedded";
 
 export const getServerSideProps: GetServerSideProps = async ({ query }) => {
   return {
@@ -13,8 +14,9 @@ export const getServerSideProps: GetServerSideProps = async ({ query }) => {
 };
 
 export function DatasetBrowser({ hideHeader }: { hideHeader: boolean }) {
+  const isEmbedded = useIsEmbedded();
   return (
-    <AppLayout hideHeader={hideHeader}>
+    <AppLayout hideHeader={hideHeader || isEmbedded}>
       <ConfiguratorStateProvider chartId="new" allowDefaultRedirect={false}>
         <SelectDatasetStep variant="page" />
       </ConfiguratorStateProvider>